🚧 Prettier reformat
A note about iconography: I use the 🚧 icon to mean, "I intended this as a pure refactor…there _shouldn't_ be any bugs possible in this commit. This practice helps me keep a clear dividing line between functional changes and things that just make the code easier to read…or, in this case, consistent with how my auto-formatter (Prettier) is set up in VS Code. It may make my commit history seem kinda messy, but I'd rather have lots of easy-to-read commits interspersed with the occasional messy-but-harmless commit like this one than have to find the functionally-relevant needle in a haystack of whitespace.