Yes! 🍻 After a few too-many hours, figured out how to smoothly animate titles and images when transitioning routes.
The biggest sticking point was that I wanted to be able to (for this first pass), keep the word "Ouroboros" on-screen when transitioning back-and-forth between the "projects" page and `routes/ouroboros/index.svelte`. The key seems to be making sure that every component involved in animations gets its _entire_ contents wrapped in a `div.route-animation-container`. Maybe this is why @buhrmi chose <main> as the place to hook `position:absolute` in the example I used to work this out: https://github.com/sveltejs/sapper/issues/778

Should I follow suit? TBD.