🚧 changed class name: `positioner` 👉 `shrinkwrap`
The latter is, I hope, more descriptive of why this div's important. In case that's not clear: it's because the `width:fit-content` wrapper prevents the animated elements from temporarily having extra-wide bounding boxes (which causes the animation to stretch the text incorrectly as the transition doesn't take into account the actual final visible size of the element, but the size it calculated based on a box that automaticaly stretches to the edge of the parent div.