`esc` to re-hide the Lorem text
This is (mostly) just for future testing ease.

(I shoulda put it in place _yesterday_ to make my manual test cycles go faster. Or maybe start taking advantage of a test suite? 😬 One-of-these-days, for sure, really! Not feeling in the mood for figuring out how to run tests in Sapper just now…too much other learning going on.)