Added super-subtle rotation on grayscale dragons
The GSAP code (`TweenMax.to…`) is pretty straightforward:
- Select the #logo
- For 360 seconds (6 min)…
  - Rotate counter-clockwise a full 360
  - No need to ease in
  - Spin indefinitely (-1)

The whole thing has to run inside Svelte's `afterUpdate` callback to force `gsap` to "see" the grayscale logo at its _final_ CSS position & scale, rather snapping it back to its _starting_ position and (much larger!) scale.

And finally, there's the `on:click` link handler. When clicking back to `projects`, we want the logo to animate smoothly back to the centeer. If we leave thee transform wherever `gsap` last left it, clicking causes the logo to fly off to the "right" in whatever orientation it's been rotated to. Resetting rotation prevents that.