Better nav positioning.
(Not sure whether the `_layout.svelte` change makes a difference, but at this point I'd rather introduce a bug that forces me to learn why a layout instruction is present than to live with the vague uneasiness that comes with unclear style choices.

The big `margin-bottom` in global.css lets content clear the fixed header.