Cleanest-looking Nav transition I've been able to work out.
That old `crossfade` transition really wants the source and destination elements to be of essentially the same type, and ESPECIALLY the same number of items, i.e., ONE. At least, that's what I've gathered after fighting this for hours. It's not exactly what I had in mind, but it'll do to be getting on with. Yeesh!

- This approach is far simpler than a number of other approaches I tried. And even so, doesn't behave the way I'd initially intended. What I _wanted_ to happen is to be able to rely on Sapper's `segment` values to conditionally display different Nav components according to the current route, but `segment` doesn't quite cut it here because it's not settable from below…without doing something like this store approach. ¯\_(ツ)_/¯
- Yes, there's little to differentiate `/ouroboros/_layout.svelte` from `/storyteller/_layout.svelte`, but I don't see a clean way to reduce duplication any further.