minor - variable rename
I'm having issues with the transition back to the main index after delving into one of the sub-indices. Part of the trouble with all of this is that I'm relying on Svelte for the typerwriter effect & inter-route transitions, but gsap for everything else. In order to function properly, There's going to have to be a very clear line of control, with gsap holding the reins and only `e` and `v` visible on mount.