Fixed positioning of the 4 worlds
Includes:
- Extract Component: Worlds

As part of this extraction, the worlds are positioned via CSS rather than GreenScok. Importantly, this particular change, along with the last one (`firstLaunch`) fixes the reverse animations that occur when returning to the main index from one of the worlds.

- New Layout Component: Fullscreen

The "Layout Component" rerpresents the first successful implementation of a simple discipline that, with any luck, will help me start to slay certain CSS dragons in a more or less permanent fashion. In a perfect worldk, there would be two types of component when it comes to CSS usage: Layout and Content. Just as Layout Components would be expected not to render anything visible on-screen, Content Components would be forbidden from including layout styles other than those imported by virtue of using Layout Comnponents to control layout _internal to_ the Content Component.

= Hooked into `firstLaunch` variable in order to skip out on the intro animation when returning from a world. (This current solution is almost certainly flawed, but my fingers are too sore right now for me to explain. Sorry.